+++ backEnd/crypto_wave/wallet_backEnd/Models/Dto/WalletCreateDto.cs
@@ -5,7 +5,7 @@ namespace user_wallet.Models.Dto
     public class WalletCreateDto
     {
         [Required]
-        public int CurrencyCount { get; set; }
+        public double CurrencyCount { get; set; }
         [Required]
         public string? CurrencyName { get; set; }
         [Required]
